Add a pull request template (#1)
Add a pull request template

### Motivation

To help maintain consistency in pull request descriptions, and encourage PR authors to provide enough information to reviewers upfront, let's add a PR template to the repo.

### Modifications

Added a PR template file based on the SwiftNIO one, with slight modifications (removed colons in headings, added a test plan section). Adjusted the NOTICE file to reflect that.

### Result

PR authors will be presented with the template when creating a new PR using GitHub's UI or tools that support templates.

### Test Plan

Well, this PR is already being open using the template :)


Reviewed by: simonjbeaumont

Builds:
     ✔︎ pull request validation (5.8) - Build finished. 
     ✔︎ pull request validation (5.9) - Build finished. 
     ✔︎ pull request validation (nightly) - Build finished. 
     ✔︎ pull request validation (soundness) - Build finished. 

https://github.com/apple/swift-openapi-runtime/pull/1