Fix encoding of plain text bodies (#9)
Fix encoding of plain text bodies

### Motivation

When the content type `text/plain` is used, we had a bug where we were encoding the text using `Codable` into a JSON fragment, i.e. `hello` -> `"hello"` (note the added quotes), which is incorrect. We should be using `LosslessStringConvertible` for primitive types, so that `hello` is sent as `hello` (without the quotes).

### Modifications

Fixed the body conversion methods to use the string conversion if possible, and only use the `Codable` implementation for complex types.

### Result

When content type is `text/plain` (and other plain text variants), `hello` is sent as `hello`, not `"hello"`.

### Test Plan

Updated unit tests and added new ones to cover these cases.


Reviewed by: simonjbeaumont

Builds:
     ✔︎ pull request validation (5.8) - Build finished. 
     ✔︎ pull request validation (5.9) - Build finished. 
     ✔︎ pull request validation (nightly) - Build finished. 
     ✔︎ pull request validation (soundness) - Build finished. 

https://github.com/apple/swift-openapi-runtime/pull/9