Rename _StringParameterConvertible to _StringConvertible (#13)
Rename _StringParameterConvertible to _StringConvertible

### Motivation

As part of reworking https://github.com/apple/swift-openapi-runtime/pull/12, we'll be using the protocol `_StringParameterConvertible` not just for parameters, but also for bodies, so renaming it accordingly to reflect that.

### Modifications

Renamed `_StringParameterConvertible` to `_StringConvertible`.

### Result

The naming is more accurate now.

Compatibility-wise, this isn't affecting the generator, as this protocol is not used directly by generated code, only indirectly through `_AutoLosslessStringConvertible`, so we don't even need to stage this change.

### Test Plan

All tests pass.


Reviewed by: simonjbeaumont

Builds:
     ✔︎ pull request validation (5.8) - Build finished. 
     ✔︎ pull request validation (5.9) - Build finished. 
     ✔︎ pull request validation (nightly) - Build finished. 
     ✔︎ pull request validation (soundness) - Build finished. 

https://github.com/apple/swift-openapi-runtime/pull/13