[Fix] Fix two regressions in Converter (#16)
[Fix] Fix two regressions in Converter

### Motivation

Found two regressions with the signature of two of the ~50 converter helper methods:
- optional request body of type data
- required header fields of type JSON

### Modifications

Fixed up these signatures.

### Result

Now a project which has both can build.

### Test Plan

Used a test project for finding and fixing these. We should think about increasing the integration test coverage, and run it on runtime PRs.


Reviewed by: simonjbeaumont

Builds:
     ✔︎ pull request validation (5.8) - Build finished. 
     ✔︎ pull request validation (5.9) - Build finished. 
     ✔︎ pull request validation (nightly) - Build finished. 
     ✔︎ pull request validation (soundness) - Build finished. 

https://github.com/apple/swift-openapi-runtime/pull/16