Converter SPI methods for required fields should return non-optional types (#20)
### Motivation

`getRequiredRequestBodyAsText` currently returns `C?` instead of `C`. As
the value should be required an optional in the return is causing
issues.

The errors are `Cannot convert value of type 'C?' to specified type` and
`Cannot infer contextual base in reference to member 'text'`

 ### Modifications

Change the return type of `getRequiredRequestBodyAsText`

 ### Result

Code using a required requestBody now compiles without errors.