Move DocC and integration tests into separate CI pipelines (#23)
### Motivation

Like many other projects, we have a soundness script which runs in CI.
It started out with some fast policy and style tests but has grown over
time. It currently has:

1. Check for broken symlinks
2. Check for unacceptable language
3. Check all files have a license headers
4. Run swift-format lint
5. Check DocC compiliation has no warnings
6. Run integration test

On my machine, (5) takes 11s and (6) takes 1m13s.

The soundness script should be cheap to run so as not to disclosure
running it often—potentially even adding it to a precommit hook.

Removing (5) and (6) from the soundness suite means it takes 3s, which
is more reasonable.

### Modifications

- Add `integration-test` to Compose file.
- Add `docc-test` to Compose file.
- Stop running DocC and integration tests in `soundness.sh`.

### Result

Running `soundness.sh` takes 3s (was ~1m30s).

### Test Plan

The following commands all succeed when run locally:

- `./scripts/soundness.sh`
- `docker-compose -f docker/docker-compose.yaml run soundness`
- `docker-compose -f docker/docker-compose.yaml run docc-test`
- `docker-compose -f docker/docker-compose.yaml run integration-test`

### Notes

This PR should be used to stand up the new CI pipelines and shouldn't be
merged
until we see them passing.