Improved Encoding/Decoding error descriptions (#28)
Improved Encoding/Decoding error descriptions

### Motivation

When an EncodingError/DecodingError is thrown and printed, we used to omit the the coding path and underlying error, which are often crucial in finding e.g. which part of the payload has a missing field.

### Modifications

This PR improves the logging to include the coding path and underlying error in the output.

### Result

Now when e.g. a DecodingError is thrown and printed, the default output has enough information to know which part of the JSON payload is malformed.

### Test Plan

Tested manually, as we don't have unit tests for exact printing strings.


Reviewed by: simonjbeaumont

Builds:
     ✔︎ pull request validation (5.8) - Build finished. 
     ✔︎ pull request validation (5.9) - Build finished. 
     ✔︎ pull request validation (api breakage) - Build finished. 
     ✔︎ pull request validation (docc test) - Build finished. 
     ✔︎ pull request validation (integration test) - Build finished. 
     ✔︎ pull request validation (nightly) - Build finished. 
     ✔︎ pull request validation (soundness) - Build finished. 

https://github.com/apple/swift-openapi-runtime/pull/28