Support url form encoded bodies (#53)
### Motivation

[Issue 182](https://github.com/apple/swift-openapi-generator/issues/182)
Supporting runtime update for [add URL form encoder &
decoder](https://github.com/apple/swift-openapi-generator/pull/283)

### Modifications

Add converter methods for Codable <--> URLEncodedForm.

**Note:** Trying to encode a primitive collection (e.g. an array of
`String`) form a Codable type fails because the URIEncoder rejects
nested object types. I didn't make any changes to the Encoder here, as I
think there may be a larger discussion needed about how to approach
this. It would be beneficial to support this, as there may be a form
which contains multiple values for the same key.

Decoding an exploded key from to a Codable type containing an array
works already, but I've not included the test in this PR, as that should
be included in the above.

### Result

Converter methods for URLEncodedForms are now available for use in
generated content.

### Test Plan

Update tests for encoders and decoders.

---------

Co-authored-by: benfrearson <ben.frearson@sonin.agency>
Co-authored-by: bfrearson <>
Co-authored-by: Honza Dvorsky <honza@apple.com>