benchmarks: Add subdirectory Benchmarks/ package (#119)
### Motivation

We'd like to be able to benchmark various parts of the runtime library
to make targeted and measurable performance improvements.

### Modifications

- Add a subdirectory directory package with the package-benchmark
plugin.
- Add a benchmark for `ISO8601DateTranscoder.encode(_:)`.
- Add temporary benchmarks of Foundation ISO8601 date encoding APIs to
guide improvements to `ISO8601DateTranscoder`.

### Result

- Can now run benchmarks using `swift package --package-path Benchmarks/
benchmark`.

### Test Plan

There is no CI for this currently. That can follow, but we wanted the
package plugin to unblock incoming updates to the date transcoder.