Support nested arrays of primitive values inside of objects (#120)
### Motivation

It's a useful pattern to define a single JSON schema for all your (e.g.
query) parameters, and handle them as a single object in your code.

In OpenAPI, that'd be expressed like this, for example:

```yaml
# parameter
name: myParams
in: query
explode: true
style: form
schema:
  $ref: '#/components/schemas/QueryObject'

# schema
QueryObject:
  type: object
  properties:
    myString:
      type: string
    myList:
      type: array
      items:
      	type: string
```

Until now, the `myList` property would not be allowed, and would fail to
serialize and parse, as arrays within objects were not allowed for
`form` style parameters (used by query items, by default).

### Modifications

This PR extends the support of the `form` style to handle single nesting
in the top level objects. It does _not_ add support for arbitrarily deep
nesting.

As part of this work, we also now allow the `deepObject` style to do the
same - use arrays nested in an object.

### Result

The useful pattern of having an array within a "params" object works
correctly now.

### Test Plan

Added unit tests for all 4 components: encoder, decoder, serializer, and
parser.