Update URIValueFromNodeDecoder for DeepObject.
### Motivation

As discussed in [[Generator] Add support of deepObject style in query params #538](https://github.com/apple/swift-openapi-generator/pull/538#discussion_r1817872441), there is a misimplementation of the `decodeRootIfPresent` method in `URIValueFromNodeDecoder`. When using the `deepObject` style, the node for `rootValue` is correctly empty, containing only the sub-objects.

Without this PR, the tests for [Add support of deepObject style in query params #538](https://github.com/apple/swift-openapi-generator/pull/538) are failing.

### Modifications

- Updated the `decodeRootIfPresent(_ type:) throws -> T` method to ignore whether `currentElementAsArray` is empty or not.

### Result

- Enables the tests in the [Generator part of the PR](https://github.com/apple/swift-openapi-generator/pull/538) to pass successfully.

### Test Plan

This PR includes unit tests that validate the change to `decodeRootIfPresent(_ type:) throws -> T` within `Test_Converter+Server` as well as in `Test_serverConversionExtensions`.